+++ changelog.md
@@ -1,6 +1,12 @@
 ### ?? v0.9.6
 
 * Fixed Gym environments changing actions list in-place during call to `step`
+* Added info on game winner to ending state
+* Added rotation matrix of players' groins to game state
+* Added velocities of body parts to game state
+* Added "get_normalized_locations" function to game state
+
+* FPS decreased by 10-20% on Linux workstation with 1000 mathframes and 2 turnframes.
 
 ### 20.8.2018 v0.9.5